LayoutTests/js/array-iterators-expected.txt

This test checks the behavior of the iterator methods on Array objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'values' in [] is true
PASS value is testArray[i]
PASS value is 1
PASS value is testArray[i]
PASS value is 2
PASS value is testArray[i]
PASS value is 3
PASS value is testArray[i]
PASS value is 4
PASS value is testArray[i]
PASS value is 5
PASS value is testArray[i]
PASS value is 6
PASS testArray.length is 6
PASS value is testArray[i]
PASS value is testArray[i]
PASS value is testArray[i]
PASS value is testArray[i]
PASS value is testArray[i]
PASS value is testArray[i]
PASS value is testArray[i]
PASS value is testArray[i]
PASS value is testArray[i]
PASS testArray.length is 9
PASS i is 0
PASS value is o[i]
PASS value is 1
PASS value is o[i]
PASS value is 2
PASS value is o[i]
PASS value is 3
PASS value is o[i]
PASS value is 4
PASS value is o[i]
PASS value is 5
PASS value is o[i]
PASS value is 6
PASS o.length is 6
PASS key is 0
PASS key is 1
PASS key is 2
PASS key is 3
PASS key is 4
PASS key is 5
PASS testArray.length is 6
PASS value is testArray[key]
PASS key is 0
PASS value is 1
PASS value is testArray[key]
PASS key is 1
PASS value is 2
PASS value is testArray[key]
PASS key is 2
PASS value is 3
PASS value is testArray[key]
PASS key is 3
PASS value is 4
PASS value is testArray[key]
PASS key is 4
PASS value is 5
PASS value is testArray[key]
PASS key is 5
PASS value is 6
PASS testArray.length is 6
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS testArray.length is 9
PASS i is 0
PASS value is o[key]
PASS key is 0
PASS value is 1
PASS value is o[key]
PASS key is 1
PASS value is 2
PASS value is o[key]
PASS key is 2
PASS value is 3
PASS value is o[key]
PASS key is 3
PASS value is 4
PASS value is o[key]
PASS key is 4
PASS value is 5
PASS value is o[key]
PASS key is 5
PASS value is 6
PASS o.length is 6
PASS value is 1.5
PASS value is 2.5
PASS value is 3.5
PASS value is 4.5
PASS value is 5.5
PASS value is 6.5
PASS testArray.length is 6
PASS value is true
PASS value is true
PASS value is true
PASS value is true
PASS value is true
PASS value is true
PASS testArray.length is 6
PASS isNaN(value) is true
PASS isNaN(value) is true
PASS isNaN(value) is true
PASS isNaN(value) is true
PASS isNaN(value) is true
PASS isNaN(value) is true
PASS testArray.length is 6
PASS value is undefined.
PASS value is undefined.
PASS value is undefined.
PASS value is undefined.
PASS value is undefined.
PASS value is undefined.
PASS value is undefined.
PASS value is undefined.
PASS testArray.length is 8
PASS key is 0
PASS key is 1
PASS key is 2
PASS key is 3
PASS key is 4
PASS key is 5
PASS key is 6
PASS key is 7
PASS testArray.length is 8
PASS successfullyParsed is true

TEST COMPLETE
